Added support for automatic scaling of photos/videos and new image/video picker in iOS14 #siskinim-247

+++ SiskinIM/chat/BaseChatViewController+ShareMedia.swift
@@ -20,13 +20,42 @@
 //
 
 import UIKit
+import PhotosUI
 
 extension ChatViewInputBar {
         
-    class ShareImageButton: ShareButton {
+    class LongPressShareButton: ShareButton {
+        
+        @objc func longPress(gesture: UILongPressGestureRecognizer) {
+            
+        }
+        
+        override func setup() {
+            super.setup();
+            let gesture = UILongPressGestureRecognizer(target: self, action: #selector(longPress(gesture:)));
+            gesture.minimumPressDuration = 1.0;
+            self.addGestureRecognizer(gesture);
+        }
+        
+    }
+    
+    class ShareImageButton: LongPressShareButton {
         
         override func execute(_ sender: Any) {
-            controller?.selectPhoto(.photoLibrary)
+            if #available(iOS 14.0, *) {
+                controller?.selectPhotoFromLibrary();
+            } else {
+                controller?.selectPhoto(.photoLibrary)
+            }
+        }
+        
+        override func longPress(gesture: UILongPressGestureRecognizer) {
+            controller?.askMediaQuality = true;
+            if #available(iOS 14.0, *) {
+                controller?.selectPhotoFromLibrary();
+            } else {
+                controller?.selectPhoto(.photoLibrary)
+            }
         }
         
         override func setup() {
@@ -40,10 +69,15 @@ extension ChatViewInputBar {
         }
     }
     
-    class ShareCameraImageButton: ShareButton {
+    class ShareCameraImageButton: LongPressShareButton {
         
         override func execute(_ sender: Any) {
             controller?.selectPhoto(.camera)
+        }
+        
+        override func longPress(gesture: UILongPressGestureRecognizer) {
+            controller?.askMediaQuality = true;
+            controller?.selectPhoto(.camera);
         }
         
         override func setup() {
@@ -57,6 +91,74 @@ extension ChatViewInputBar {
         }
     }
 
+}
+
+
+@available(iOS 14.0, *)
+extension BaseChatViewController: PHPickerViewControllerDelegate {
+        
+    func selectPhotoFromLibrary() {
+        var config = PHPickerConfiguration();
+        config.selectionLimit = 1;
+        config.filter = .any(of: [.videos, .images]);
+        config.preferredAssetRepresentationMode = .current;
+                
+        let picker = PHPickerViewController(configuration: config);
+        picker.delegate = self;
+        
+        present(picker, animated: true);
+    }
+    
+    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
+        picker.dismiss(animated: true);
+        
+        if let provider = results.first?.itemProvider {
+            if provider.canLoadObject(ofClass: UIImage.self) {
+                provider.loadFileRepresentation(forTypeIdentifier: "public.image", completionHandler: self.handleLoaded(imageUrl:error:));
+            } else if provider.hasItemConformingToTypeIdentifier("public.movie") {
+                provider.loadFileRepresentation(forTypeIdentifier: "public.movie", completionHandler: self.handleLoaded(movieUrl:error:));
+            } else {
+                showAlert(shareError: .notSupported);
+            }
+        }
+    }
+    
+    private func handleLoaded(imageUrl url: URL?, error: Error?) {
+        guard let url = url, error == nil else {
+            DispatchQueue.main.async {
+                self.showAlert(shareError: .noAccessError);
+            }
+            return;
+        }
+
+        guard let (localUrl, filename, _) = copyFileLocally(url: url, defaultSuffix: "jpg") else {
+            DispatchQueue.main.async {
+                self.showAlert(shareError: .unknownError);
+            }
+            return;
+        }
+
+        upload(imageUrl: localUrl, filename: filename);
+    }
+    
+    private func handleLoaded(movieUrl url: URL?, error: Error?) {
+        guard let url = url, error == nil else {
+            DispatchQueue.main.async {
+                self.showAlert(shareError: .noAccessError);
+            }
+            return;
+        }
+        
+        guard let (localUrl, filename, _) = copyFileLocally(url: url, defaultSuffix: "mov") else {
+            DispatchQueue.main.async {
+                self.showAlert(shareError: .unknownError);
+            }
+            return;
+        }
+
+        upload(movieUrl: localUrl, filename: filename);
+    }
+    
 }
 
 extension BaseChatViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
@@ -78,14 +180,100 @@ extension BaseChatViewController: UIImagePickerControllerDelegate, UINavigationC
     }
     
     @objc func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        guard let fileUrl = (info[UIImagePickerController.InfoKey.mediaURL] as? URL) ?? (info[UIImagePickerController.InfoKey.imageURL] as? URL) else {
-            return;
+        if let movieUrl = info[.mediaURL] as? URL {
+            var filename = movieUrl.lastPathComponent;
+            if let idx = filename.lastIndex(of: ".") {
+                filename = String(filename.prefix(upTo: idx));
+                if filename.hasPrefix("trim.") {
+                    filename = String(filename.dropFirst("trim.".count));
+                }
+            }
+
+            upload(movieUrl: movieUrl, filename: filename);
+        } else if let imageUrl = info[.imageURL] as? URL {
+            var filename = imageUrl.lastPathComponent;
+            if let idx = filename.lastIndex(of: ".") {
+                filename = String(filename.prefix(upTo: idx));
+                if filename.hasPrefix("trim.") {
+                    filename = String(filename.dropFirst("trim.".count));
+                }
+            }
+
+            upload(imageUrl: imageUrl, filename: filename);
         }
         
-        var filename = fileUrl.lastPathComponent;
-        if filename.hasPrefix("trim.") {
-            filename = String(filename.dropFirst("trim.".count));
+        picker.dismiss(animated: true, completion: nil);
+    }
+    
+    func upload(imageUrl url: URL, filename: String) {
+        MediaHelper.askImageQuality(controller: self, forceQualityQuestion: self.askMediaQuality, { result in
+            self.askMediaQuality = false;
+            switch result {
+            case .success(let quality):
+                MediaHelper.compressImage(url: url, filename: filename, quality: quality, deleteSource: true, completionHandler: { result in
+                    switch result {
+                    case .success(let fileUrl):
+                        self.uploadFile(url: fileUrl, filename: fileUrl.lastPathComponent, deleteSource: true);
+                    case .failure(let error):
+                        self.showAlert(shareError: error);
+                    }
+                })
+            case .failure(_):
+                return;
+            }
+        });
+    }
+    
+    func upload(movieUrl url: URL, filename: String) {
+        MediaHelper.askVideoQuality(controller: self, forceQualityQuestion: self.askMediaQuality, { result in
+            self.askMediaQuality = false;
+            switch result {
+            case .success(let quality):
+                DispatchQueue.main.async {
+                    self.showProgressBar();
+                    MediaHelper.compressMovie(url: url, filename: filename, quality: quality, deleteSource: true, progressCallback: { [weak self] progress in
+                        DispatchQueue.main.async {
+                            self?.progressBar?.progress = progress;
+                        }
+                    }, completionHandler: { result in
+                        DispatchQueue.main.async {
+                            self.hideProgressBar();
+                        }
+                        switch result {
+                        case .success(let fileUrl):
+                            self.uploadFile(url: fileUrl, filename: fileUrl.lastPathComponent, deleteSource: true);
+                        case .failure(let error):
+                            self.showAlert(shareError: error);
+                        }
+                    })
+                }
+            case .failure(_):
+                return;
+            }
+        });
+    }
+    
+    private func copyFileLocally(url: URL, defaultSuffix: String) -> (URL, String, String)? {
+        var filename = url.lastPathComponent;
+        var suffix = defaultSuffix;
+        if let idx = filename.lastIndex(of: ".") {
+            suffix = String(filename.suffix(from: filename.index(after: idx)));
+            filename = String(filename.prefix(upTo: idx));
+            if filename.hasPrefix("trim.") {
+                filename = String(filename.dropFirst("trim.".count));
+            }
         }
+        
+        let tmpUrl = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString + "." + suffix, isDirectory: false);
+        do {
+            try FileManager.default.copyItem(at: url, to: tmpUrl);
+        } catch {
+            return nil;
+        }
+        return (tmpUrl, filename, suffix);
+    }
+        
+    private func uploadFile(url fileUrl: URL, filename: String, deleteSource: Bool) {
         self.share(filename: filename, url: fileUrl, completionHandler: { result in
             switch result {
             case .success(let uploadedUrl, let filesize, let mimetype):
@@ -97,7 +285,7 @@ extension BaseChatViewController: UIImagePickerControllerDelegate, UINavigationC
                 appendix.state = .downloaded;
 
                 self.sendAttachment(originalUrl: fileUrl, uploadedUrl: uploadedUrl.absoluteString, appendix: appendix, completionHandler: {
-                    if FileManager.default.fileExists(atPath: fileUrl.path) {
+                    if deleteSource && FileManager.default.fileExists(atPath: fileUrl.path) {
                         try? FileManager.default.removeItem(at: fileUrl);
                     }
                 });
@@ -105,8 +293,7 @@ extension BaseChatViewController: UIImagePickerControllerDelegate, UINavigationC
                 self.showAlert(shareError: error);
             }
         })
-        picker.dismiss(animated: true, completion: nil);
     }
-
+    
 }
